+++ src/main/java/io/github/mzmine/modules/visualization/rawdataoverviewims/IMSRawDataOverviewTab.java
@@ -46,7 +46,7 @@ public class IMSRawDataOverviewTab extends MZmineTab {
         parameterSet.getParameter(IMSRawDataOverviewParameters.mzTolerance).getValue(),
         parameterSet.getParameter(IMSRawDataOverviewParameters.scanSelection).getValue(),
         parameterSet.getParameter(IMSRawDataOverviewParameters.rtWidth).getValue().floatValue(),
-        parameterSet.getParameter(IMSRawDataOverviewParameters.binWidth).getValue().doubleValue());
+        parameterSet.getParameter(IMSRawDataOverviewParameters.binWidth).getValue());
     setContent(pane);
   }
